Time is now dealt in seconds rather than nanosecs

+++ inc/timer.h
@@ -34,6 +34,6 @@ struct chess_timer* update_timer(struct chess_timer* timer, struct chess* game);
 /**
  * Returns the remaining time for a specific move.
  */
-long get_remaining_move_time(struct chess_timer* timer);
+double get_remaining_move_time(struct chess_timer* timer);
 
 #endif /* TIMER_H */
